+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - import fc from gltf / RotationY
 - empty line in .cf7 endless loop in annofcconverter
 - no tangent warning in channel output
+- error handling for invalid xml files
 
 ### Added
 
@@ -34,6 +35,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Don't overwrite files with `Convert to ..` context menu commands
 - Hover info for asset keywords like `UpgradeList` and `ResidenceUpgrade`
 - Auto-complete after typing `GUID='`
+- Import `FILE` sections with `Import PROPs from glTF`.
 
 ## [0.2.0]